chore: move to the new dir structure
