+++ manifest.toml
@@ -148,3 +148,19 @@ instances = { min = 4, max = 3000, default = 12 }
     light = { type = "int", default = 3}
     block-height = { type = "int" }
     role = { type = "string" }
+
+[[testcases]]
+name = "001-lights-dasing-latest"
+instances = { min = 30, max = 200, default = 30 }
+    [testcases.params]
+    execution-time = { type = "int" }
+    latency = { type = "int", default = 60}
+    bandwidth = { type = "string", default = "256Mib"}
+    validator = { type = "int", default = 3}
+    persistent-peers = { type = "int", default = 3}
+    submit-times = { type = "int", default = 4}
+    msg-size = { type = "int", default = 10000 }
+    bridge = { type = "int", default = 3}
+    light = { type = "int", default = 3}
+    block-height = { type = "int" }
+    role = { type = "string" }
