--- plans/002-das-benchmarks/README.md
@@ -1,8 +1,0 @@
-# Tests
-
-Currently, to test the DASing behavior with `SampleAmount=50`, we are relying on a forked version of celestia node that hard codes this configuration value (_because this value is not configureable yet_)
-
-The version we are currently pointing at is:
-```
-github.com/derrandz/celestia-node v0.4.2-0.20221025143831-b31eaa3d9cdb
-```
