--- plans/002-das-benchmarks/manifest.toml
@@ -1,37 +1,0 @@
-name = "002-das-benchmarks"
-
-[defaults]
-builder = "docker:generic"
-runner = "local:exec"
-
-[builders."docker:generic"]
-enabled = true
-build_base_image = "golang:1.19.1"
-go_version = "1.19"
-enable_go_build_cache = true
-
-[runners."local:docker"]
-enabled = true
-
-[runners."local:exec"]
-enabled = true
-
-[runners."cluster:k8s"]
-enabled = true
-
-
-[[testcases]]
-name = "001-lights-dasing-latest"
-instances = { min = 30, max = 200, default = 30 }
-    [testcases.params]
-    execution-time = { type = "int" }
-    latency = { type = "int", default = 60}
-    bandwidth = { type = "string", default = "256Mib"}
-    validator = { type = "int", default = 3}
-    persistent-peers = { type = "int", default = 3}
-    submit-times = { type = "int", default = 4}
-    msg-size = { type = "int", default = 10000 }
-    bridge = { type = "int", default = 3}
-    light = { type = "int", default = 3}
-    block-height = { type = "int" }
-    role = { type = "string" }
