+++ tests/helpers/common/bridge.go
@@ -14,6 +14,7 @@ import (
 	"github.com/testground/sdk-go/run"
 	"github.com/testground/sdk-go/runtime"
 	"github.com/testground/sdk-go/sync"
+	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp"
 )
 
 func BuildBridge(ctx context.Context, runenv *runtime.RunEnv, initCtx *run.InitContext) (*nodebuilder.Node, error) {
@@ -45,7 +46,19 @@ func BuildBridge(ctx context.Context, runenv *runtime.RunEnv, initCtx *run.InitC
 	cfg.Gateway.Enabled = true
 	cfg.Gateway.Port = "26659"
 
-	nd, err := nodekit.NewNode(ndhome, node.Bridge, cfg)
+	nd, err := nodekit.NewNode(
+		ndhome,
+		node.Bridge,
+		cfg,
+		nodebuilder.WithMetrics(
+			[]otlpmetrichttp.Option{
+				otlpmetrichttp.WithEndpoint("192.18.0.8:4318"),
+				otlpmetrichttp.WithInsecure(),
+			},
+			nodekit.BridgeNodeType,
+		),
+	)
+
 	if err != nil {
 		return nil, err
 	}
