+++ tests/helpers/das-benchmarks/run_bridge.go
@@ -1,4 +1,4 @@
-package synclatest
+package dasbenchmarks
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 
 	"github.com/celestiaorg/test-infra/testkit"
 	"github.com/celestiaorg/test-infra/testkit/nodekit"
-	"github.com/celestiaorg/test-infra/testkit/common"
+	"github.com/celestiaorg/test-infra/tests/helpers/common"
 	"github.com/testground/sdk-go/network"
 	"github.com/testground/sdk-go/run"
 	"github.com/testground/sdk-go/runtime"
